demo/phantom/client/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# re/client.py: 10
PyInquirer >= 1.0.3

# re/client.py: 21,22
argon2_cffi >= 19.2.0

# re/client.py: 24
colorama >= 0.4.3

# re/client.py: 14,18
cryptography >= 2.9

# re/client.py: 13
ipaddress >= 1.0.23

# re/client.py: 15
pycrypto >= 2.6.1

# re/client.py: 7
pynput >= 1.6.8

# re/client.py: 17
simple_crypt >= 4.1.7
